styled login and signup pages
