+++ src/components/Dashboard.js
@@ -5,6 +5,7 @@ import { useState, useEffect } from 'react';
 // database
 import firebase from '../firebase';
 import { getDatabase, ref, onValue, remove } from 'firebase/database'
+import { Link } from 'react-router-dom';
 
 function Dashboard() {
 
@@ -86,7 +87,20 @@ function Dashboard() {
                         />
                     })
                 }  
-            </div>          
+            </div>
+
+            {/* Credits */}
+            <div className="siteCredits">
+                <p className="credits">
+                    Images provided by <a className='footerLink' href="https://unsplash.com/developers"> Unsplash API</a> and profanity filter by <a className='footerLink' href="https://www.purgomalum.com/index.html">PurgoMalum RESTful API.</a>
+                </p>
+                <p className="credits">
+                    Header background provided by <a className='footerLink' href="https://loading.io/"> loading.io</a>
+                </p>
+                <Link to={`/error`}>
+                    <p className='footerLink'>Don't click this 👀</p>
+                </Link> 
+            </div>
         </section>
     )
 }
